act/requirements.txt
Flask
requests
psycopg2-binary
Flask-JWT-Extended
grpcio
grpcio-tools